requirements.txt
certifi==2019.9.11
cffi==1.13.1
cryptography==2.8
emoji==0.5.4
future==0.18.1
pycparser==2.19
PySocks==1.7.1
python-telegram-bot==10.1.0
six==1.12.0
tornado==6.0.3
clarifai
